Finish endpoint logic and test
