+++ src/db/client/async_.py
@@ -59,6 +59,8 @@ from src.db.models.instantiations.url.suggestion.relevant.auto import AutoReleva
 from src.db.models.instantiations.url.suggestion.relevant.user import UserRelevantSuggestion
 from src.db.queries.implementations.core.final_review.get import GetNextURLForFinalReviewQueryBuilder
 from src.db.queries.implementations.core.get_recent_batch_summaries.builder import GetRecentBatchSummariesQueryBuilder
+from src.db.queries.implementations.core.metrics.urls.aggregated.pending import \
+    GetMetricsURLSAggregatedPendingQueryBuilder
 from src.db.statement_composer import StatementComposer
 from src.db.constants import PLACEHOLDER_AGENCY_NAME
 from src.db.enums import TaskType
@@ -2284,5 +2286,13 @@ class AsyncDatabaseClient:
             total_urls=annotated_result + not_annotated_result
         )
 
-    async def get_urls_aggregated_pending_metrics(self):
-        pass
+    @session_manager
+    async def get_urls_aggregated_pending_metrics(
+        self,
+        session: AsyncSession
+    ):
+        builder = GetMetricsURLSAggregatedPendingQueryBuilder()
+        result = await builder.run(
+            session=session
+        )
+        return result
