+++ tests/helpers/api_test_helper.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 from src.api.endpoints.batch.dtos.get.summaries.response import GetBatchSummariesResponse
 from src.core.core import AsyncCore
 from src.core.enums import BatchStatus
-from tests.automated.integration.api.helpers.RequestValidator import RequestValidator
+from tests.automated.integration.api._helpers.RequestValidator import RequestValidator
 from tests.helpers.db_data_creator import DBDataCreator
